Login ecommerce

+++ PDOMySQL/sistema2.0/db_sistema.sql
@@ -2,7 +2,7 @@ DROP TABLE IF EXISTS `usuario`;
 CREATE TABLE IF NOT EXISTS `usuario` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `nombre` varchar(50) NOT NULL,
-  `telefono` bigint(20) NOT NULL,
+  `telefono` int(20) NOT NULL,
   `email` varchar(60) NOT NULL,
   PRIMARY KEY (`id`)
 );
